Capítulo 4 - Mapeamento

+++ src/main/java/br/com/alura/jpa/modelo/Conta.java
@@ -4,6 +4,7 @@ import javax.persistence.Entity;
 import javax.persistence.GeneratedValue;
 import javax.persistence.GenerationType;
 import javax.persistence.Id;
+import java.math.BigDecimal;
 
 @Entity
 public class Conta {
@@ -13,9 +14,9 @@ public class Conta {
     private String titular;
     private Integer agencia;
     private Integer numero;
-    private Double saldo;
+    private BigDecimal saldo;
 
-    public Conta(String titular, Integer agencia, Integer numero, Double saldo) {
+    public Conta(String titular, Integer agencia, Integer numero, BigDecimal saldo) {
         this.titular = titular;
         this.agencia = agencia;
         this.numero = numero;
@@ -52,11 +53,11 @@ public class Conta {
         this.numero = numero;
     }
 
-    public Double getSaldo() {
+    public BigDecimal getSaldo() {
         return saldo;
     }
 
-    public void setSaldo(Double saldo) {
+    public void setSaldo(BigDecimal saldo) {
         this.saldo = saldo;
     }
 }
